Move `ChainWorkerStateWithTemporaryChanges` type
Place it in its own module.
